Began adding state machine logic to the keyboard module.
To decode keys it is required that a state machine monitors input
packets to mark the special control packets that can come from the
keyboard. To do this, a state machine is used to check incoming packets
and set special output bits that are only latched when a new key press
has been completely received from the keyboard. Once this output is
available the key press can be deocded to an ASCII character.